Optimizes memory usage for extra lengths
Reduces memory footprint by changing the data type of the `buf_extras` buffer from `uint32_t*` to `uint8_t*`.

This change also impacts the calculation of `sz_extras` in `zxc_cctx_init` to correctly allocate enough memory for variable-length bytes.

Updates the encoding logic in `zxc_encode_block_gnr` and `zxc_encode_block_rec` to use `zxc_write_vbyte` for writing variable-length bytes directly to the `buf_extras` buffer and adapts the block descriptor size accounting for the variable byte size.

+++ include/zxc_sans_io.h
@@ -67,7 +67,7 @@ typedef struct {
     uint32_t* buf_sequences;  // Buffer for sequence records (packed: LL (8) | ML (8) | Offset (16))
     uint8_t* buf_tokens;      // Buffer for token sequences
     uint16_t* buf_offsets;    // Buffer for offsets
-    uint32_t* buf_extras;     // Buffer for extra lengths (vbytes for LL/ML)
+    uint8_t* buf_extras;      // Buffer for extra lengths (vbytes for LL/ML)
     uint8_t* literals;        // Buffer for literal bytes
 
     // Cold zone: configuration / scratch / resizeable
